+++ src/main/java/com/eight/sailingship/entity/Menu.java
@@ -36,11 +36,6 @@ public class Menu {
     @OneToMany(mappedBy = "menu", cascade = {CascadeType.PERSIST, CascadeType.REMOVE}, orphanRemoval = true)
     private List<OrderMenu> orderMenuList;
 
-    public void addOrderMenuList(OrderMenu orderMenu) {
-        this.orderMenuList.add(orderMenu);
-        orderMenu.setMenu(this);
-    }
-
     public void update(MenuRequestDto requestDto) {
         this.menuName = requestDto.getMenuName();
         this.introduce = requestDto.getIntroduce();
